uORB: factor out ORBMap.hpp into a separate file
The new uORB::ORBMap class was put it its own file with proper
copyright.
